web/crawl4ai/pyproject.toml
[tool.poetry]
name = "crawl4ai"
version = "0.1.0"
description = ""
authors = ["wwt2008 <wwt2008@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
